# Conflicts:
#	app/build.gradle.kts
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/sesecoffee/AddProductActivity.kt
#	app/src/main/java/com/example/sesecoffee/AdminMainActivity.kt
#	app/src/main/java/com/example/sesecoffee/CartOrderActivity.kt
#	app/src/main/java/com/example/sesecoffee/EditProductActivity.kt
#	app/src/main/java/com/example/sesecoffee/MainActivity.kt
#	app/src/main/java/com/example/sesecoffee/PaymentActivity.kt
#	app/src/main/java/com/example/sesecoffee/ProductOrderActivity.kt
#	app/src/main/java/com/example/sesecoffee/ProductUpdateOrderActivity.kt
#	app/src/main/java/com/example/sesecoffee/ProfileActivity.kt
#	app/src/main/java/com/example/sesecoffee/RatingActivity.kt
#	app/src/main/java/com/example/sesecoffee/RedeemPaymentActivity.kt
#	app/src/main/java/com/example/sesecoffee/adapters/OrderAdapter.kt
#	app/src/main/java/com/example/sesecoffee/adapters/RedeemAdapter.kt
#	app/src/main/java/com/example/sesecoffee/enums/PaymentMethod.kt
#	app/src/main/java/com/example/sesecoffee/fragments/AdminRedeemFragment.kt
#	app/src/main/java/com/example/sesecoffee/fragments/HomeFragment.kt
#	app/src/main/java/com/example/sesecoffee/fragments/RedeemFragment.kt
#	app/src/main/java/com/example/sesecoffee/model/Order.kt
#	app/src/main/java/com/example/sesecoffee/model/OrderItem.kt
#	app/src/main/java/com/example/sesecoffee/utils/Format.kt
#	app/src/main/res/layout/activity_add_redeem.xml
#	app/src/main/res/layout/activity_cart_order.xml
#	app/src/main/res/layout/activity_edit_product.xml
#	app/src/main/res/layout/activity_edit_redeem.xml
#	app/src/main/res/layout/activity_payment.xml
#	app/src/main/res/layout/activity_rating.xml
#	app/src/main/res/layout/activity_redeem_payment.xml
#	app/src/main/res/layout/activity_sign_in.xml
#	app/src/main/res/layout/activity_sign_up.xml
#	app/src/main/res/layout/admin_redeem_item.xml
#	app/src/main/res/layout/fragment_admin_home.xml
#	app/src/main/res/layout/fragment_admin_redeem.xml
#	app/src/main/res/layout/fragment_redeem.xml
#	app/src/main/res/layout/fragment_reward.xml
#	app/src/main/res/values/colors.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/sesecoffee/EditProductActivity.kt
@@ -1,5 +1,7 @@
 package com.example.sesecoffee
 
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.content.Intent
 import android.net.Uri
 import android.os.Build
@@ -101,10 +103,30 @@ class EditProductActivity : AppCompatActivity() {
             }
 
             productDeleteBtn.setOnClickListener(){
-                // not testing yet
-                if(oldProduct != null){
-                    oldProduct?.id?.let { it1 -> productsViewModel.deleteProduct(it1) }
+                val context = it.context
+                val alertDialog: AlertDialog? = this.let {
+                    val builder = AlertDialog.Builder(context)
+                    builder.apply {
+                        setPositiveButton("Yes", DialogInterface.OnClickListener { dialog, id ->
+                            if(oldProduct != null){
+                                oldProduct?.id?.let { it1 -> productsViewModel.deleteProduct(it1) }
+                            }
+                        })
+                        setNegativeButton("No", DialogInterface.OnClickListener { dialog, id ->
+
+                        })
+                        // Set other dialog properties
+                        setIcon(R.drawable.ic_warning_yellow)
+                        setTitle("Do you want to delete this product?")
+                    }
+                    // Create the AlertDialog
+                    builder.create()
                 }
+
+                if (alertDialog != null) {
+                    alertDialog!!.show()
+                }
+
             }
         }
     }
@@ -121,11 +143,12 @@ class EditProductActivity : AppCompatActivity() {
         if (oldProduct!= null) {
             val id = oldProduct?.id
             var proName = binding.productInputName.text.toString()
+            var proDesc = binding.productInputDescription.text.toString()
             var proPrice = binding.productInputPrice.text.toString()
             var imageUriString: String = imageUri.toString()
             var timeStamp: Timestamp = Timestamp(Date())
             var proPriceInt = if (proPrice.isNotEmpty()) proPrice.toInt() else 0
-            var product = Product(id, proName, proPriceInt, imageUriString, timeStamp)
+            var product = Product(id, proName,proDesc, proPriceInt, imageUriString, timeStamp)
 
             productsViewModel.updateProduct(product, imageUri, oldProduct?.imageUrl)
         }
